Reorganized the IT's to clearly separate the Agrona-based ones.
