+++ src/test/java/org/reaktivity/nukleus/http/internal/streams/rfc7230/client/ArchitectureIT.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package org.reaktivity.nukleus.http.internal.streams.client.rfc7230;
+package org.reaktivity.nukleus.http.internal.streams.rfc7230.client;
 
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.junit.rules.RuleChain.outerRule;
@@ -31,8 +31,8 @@ public class ArchitectureIT
 {
     private final K3poRule k3po = new K3poRule()
             .addScriptRoot("route", "org/reaktivity/specification/nukleus/http/control/route")
-            .addScriptRoot("client", "org/reaktivity/specification/nukleus/http/streams/rfc7230/architecture")
-            .addScriptRoot("server", "org/reaktivity/specification/http/rfc7230/architecture");
+            .addScriptRoot("server", "org/reaktivity/specification/http/rfc7230/architecture")
+            .addScriptRoot("client", "org/reaktivity/specification/nukleus/http/streams/rfc7230/architecture");
 
     private final TestRule timeout = new DisableOnDebug(new Timeout(5, SECONDS));
 
@@ -40,19 +40,14 @@ public class ArchitectureIT
         .directory("target/nukleus-itests")
         .commandBufferCapacity(1024)
         .responseBufferCapacity(1024)
-        .counterValuesBufferCapacity(1024)
-        .streams("http", "source")
-        .streams("source", "http#source")
-        .streams("target", "http#source")
-        .streams("http", "target")
-        .streams("source", "http#target");
+        .counterValuesBufferCapacity(1024);
 
     @Rule
     public final TestRule chain = outerRule(nukleus).around(k3po).around(timeout);
 
     @Test
     @Specification({
-        "${route}/input/new/controller",
+        "${route}/output/new/controller",
         "${client}/request.and.response/client",
         "${server}/request.and.response/server" })
     public void shouldCorrelateRequestAndResponse() throws Exception
@@ -62,7 +57,7 @@ public class ArchitectureIT
 
     @Test
     @Specification({
-        "${route}/input/new/controller",
+        "${route}/output/new/controller",
         "${client}/request.uri.with.percent.chars/client",
         "${server}/request.uri.with.percent.chars/server" })
     public void shouldAcceptRequestWithPercentChars() throws Exception
