+++ src/test/java/org/reaktivity/nukleus/http/internal/streams/rfc7230/client/ConnectionManagementIT.java
@@ -13,35 +13,28 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package org.reaktivity.nukleus.http.internal.streams.server.rfc7230;
+package org.reaktivity.nukleus.http.internal.streams.rfc7230.client;
 
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.junit.rules.RuleChain.outerRule;
-import static org.reaktivity.nukleus.http.internal.Context.MAXIMUM_HEADERS_SIZE_PROPERTY_NAME;
-import static org.reaktivity.nukleus.http.internal.Context.MEMORY_FOR_DECODE_PROPERTY_NAME;
 
+import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.DisableOnDebug;
 import org.junit.rules.TestRule;
 import org.junit.rules.Timeout;
-import org.kaazing.k3po.junit.annotation.ScriptProperty;
 import org.kaazing.k3po.junit.annotation.Specification;
 import org.kaazing.k3po.junit.rules.K3poRule;
-import org.reaktivity.nukleus.http.internal.test.SystemPropertiesRule;
 import org.reaktivity.reaktor.test.NukleusRule;
 
-public class FlowControlLimitsIT
+public class ConnectionManagementIT
 {
     private final K3poRule k3po = new K3poRule()
             .addScriptRoot("route", "org/reaktivity/specification/nukleus/http/control/route")
-            .addScriptRoot("streams", "org/reaktivity/specification/nukleus/http/streams/rfc7230/flow.control");
+            .addScriptRoot("streams", "org/reaktivity/specification/nukleus/http/streams/rfc7230/connection.management/agrona");
 
     private final TestRule timeout = new DisableOnDebug(new Timeout(5, SECONDS));
-
-    private final TestRule properties = new SystemPropertiesRule()
-        .setProperty(MAXIMUM_HEADERS_SIZE_PROPERTY_NAME, "64")
-        .setProperty(MEMORY_FOR_DECODE_PROPERTY_NAME, "64");
 
     private final NukleusRule nukleus = new NukleusRule("http")
         .directory("target/nukleus-itests")
@@ -55,13 +48,14 @@ public class FlowControlLimitsIT
         .streams("source", "http#target");
 
     @Rule
-    public final TestRule chain = outerRule(properties).around(nukleus).around(k3po).around(timeout);
+    public final TestRule chain = outerRule(nukleus).around(k3po).around(timeout);
 
     @Test
     @Specification({
         "${route}/input/new/controller",
-        "${streams}/request.headers.too.long/server/source" })
-    public void shouldRejectRequestExceedingMaximumHeadersSize() throws Exception
+        "${streams}/response.status.101.with.upgrade/server/source",
+        "${streams}/response.status.101.with.upgrade/server/target" })
+    public void shouldSwitchProtocolAfterUpgrade() throws Exception
     {
         k3po.start();
         k3po.awaitBarrier("ROUTED_INPUT");
@@ -72,46 +66,22 @@ public class FlowControlLimitsIT
     @Test
     @Specification({
         "${route}/input/new/controller",
-        "${streams}/response.headers.too.long/server/source",
-        "${streams}/response.headers.too.long/server/target" })
-    public void shouldNotWriteResponseExceedingMaximumHeadersSize() throws Exception
+        "${streams}/multiple.requests/server/source",
+        "${streams}/multiple.requests/server/target" })
+    public void shouldAcceptMultipleRequests() throws Exception
     {
         k3po.start();
         k3po.awaitBarrier("ROUTED_INPUT");
         k3po.notifyBarrier("ROUTED_OUTPUT");
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${route}/output/new/controller",
-        "${streams}/response.headers.too.long/client/source",
-        "${streams}/response.headers.too.long/client/target"})
-    public void shouldRejectResponseExceedingMaximumHeadersSize() throws Exception
-    {
-        k3po.start();
-        k3po.awaitBarrier("ROUTED_OUTPUT");
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${route}/output/new/controller",
-        "${streams}/request.headers.too.long/client/source" })
-    public void shouldNotWriteRequestExceedingMaximumHeadersSize() throws Exception
-    {
-        k3po.start();
-        k3po.awaitBarrier("ROUTED_OUTPUT");
         k3po.finish();
     }
 
     @Test
     @Specification({
         "${route}/input/new/controller",
-        "${streams}/request.fragmented.with.content.length/server/source",
-        "${streams}/request.fragmented.with.content.length/server/target" })
-    @ScriptProperty("targetInputInitialWindow [0x40 0x00 0x00 0x00]") // 64 bytes, same as max headers size
-    public void shouldAcceptFragmentedRequestWithDataWhenOnlyDataExceedsMaxHttpHeadersSize() throws Exception
+        "${streams}/multiple.requests.same.connection/server/source",
+        "${streams}/multiple.requests.same.connection/server/target" })
+    public void shouldAcceptMultipleRequestsOnSameConnection() throws Exception
     {
         k3po.start();
         k3po.awaitBarrier("ROUTED_INPUT");
@@ -119,14 +89,12 @@ public class FlowControlLimitsIT
         k3po.finish();
     }
 
-
     @Test
     @Specification({
         "${route}/output/new/controller",
-        "${streams}/response.fragmented.with.content.length/client/source",
-        "${streams}/response.fragmented.with.content.length/client/target" })
-    @ScriptProperty("sourceInputInitialWindow [0x40 0x00 0x00 0x00]") // 64 bytes, same as max headers size
-    public void shouldAcceptFragmentedResponseWithDataWhenOnlyDataExceedsMaxHttpHeadersSize() throws Exception
+        "${streams}/response.status.101.with.upgrade/client/source",
+        "${streams}/response.status.101.with.upgrade/client/target" })
+    public void shouldSwitchProtocolAfterUpgradeClient() throws Exception
     {
         k3po.start();
         k3po.awaitBarrier("ROUTED_OUTPUT");
@@ -134,5 +102,30 @@ public class FlowControlLimitsIT
         k3po.finish();
     }
 
-}
+    @Test
+    @Specification({
+        "${route}/output/new/controller",
+        "${streams}/multiple.requests/client/source",
+        "${streams}/multiple.requests/client/target" })
+    public void shouldIssueMultipleRequests() throws Exception
+    {
+        k3po.start();
+        k3po.awaitBarrier("ROUTED_OUTPUT");
+        k3po.notifyBarrier("ROUTED_INPUT");
+        k3po.finish();
+    }
 
+    @Test
+    @Specification({
+        "${route}/output/new/controller",
+        "${streams}/multiple.requests.same.connection/client/source",
+        "${streams}/multiple.requests.same.connection/client/target" })
+    @Ignore("Not yet supported")
+    public void shouldIssueMultipleRequestsUsingConnectionPool() throws Exception
+    {
+        k3po.start();
+        k3po.awaitBarrier("ROUTED_OUTPUT");
+        k3po.notifyBarrier("ROUTED_INPUT");
+        k3po.finish();
+    }
+}
