+++ src/test/java/org/reaktivity/nukleus/http/internal/streams/rfc7230/server/ConnectionManagementIT.java
@@ -13,11 +13,12 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package org.reaktivity.nukleus.http.internal.streams.server.rfc7230;
+package org.reaktivity.nukleus.http.internal.streams.rfc7230.server;
 
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.junit.rules.RuleChain.outerRule;
 
+import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.DisableOnDebug;
@@ -27,11 +28,11 @@ import org.kaazing.k3po.junit.annotation.Specification;
 import org.kaazing.k3po.junit.rules.K3poRule;
 import org.reaktivity.reaktor.test.NukleusRule;
 
-public class MessageFormatIT
+public class ConnectionManagementIT
 {
     private final K3poRule k3po = new K3poRule()
             .addScriptRoot("route", "org/reaktivity/specification/nukleus/http/control/route")
-            .addScriptRoot("streams", "org/reaktivity/specification/nukleus/http/streams/rfc7230/message.format");
+            .addScriptRoot("streams", "org/reaktivity/specification/nukleus/http/streams/rfc7230/connection.management/agrona");
 
     private final TestRule timeout = new DisableOnDebug(new Timeout(5, SECONDS));
 
@@ -52,9 +53,9 @@ public class MessageFormatIT
     @Test
     @Specification({
         "${route}/input/new/controller",
-        "${streams}/request.with.content.length/server/source",
-        "${streams}/request.with.content.length/server/target" })
-    public void shouldAcceptRequestWithContentLength() throws Exception
+        "${streams}/response.status.101.with.upgrade/server/source",
+        "${streams}/response.status.101.with.upgrade/server/target" })
+    public void shouldSwitchProtocolAfterUpgrade() throws Exception
     {
         k3po.start();
         k3po.awaitBarrier("ROUTED_INPUT");
@@ -65,9 +66,9 @@ public class MessageFormatIT
     @Test
     @Specification({
         "${route}/input/new/controller",
-        "${streams}/request.with.headers/server/source",
-        "${streams}/request.with.headers/server/target" })
-    public void shouldAcceptRequestWithHeaders() throws Exception
+        "${streams}/multiple.requests/server/source",
+        "${streams}/multiple.requests/server/target" })
+    public void shouldAcceptMultipleRequests() throws Exception
     {
         k3po.start();
         k3po.awaitBarrier("ROUTED_INPUT");
@@ -78,9 +79,9 @@ public class MessageFormatIT
     @Test
     @Specification({
         "${route}/input/new/controller",
-        "${streams}/response.with.content.length/server/source",
-        "${streams}/response.with.content.length/server/target" })
-    public void shouldWriteResponseWithContentLength() throws Exception
+        "${streams}/multiple.requests.same.connection/server/source",
+        "${streams}/multiple.requests.same.connection/server/target" })
+    public void shouldAcceptMultipleRequestsOnSameConnection() throws Exception
     {
         k3po.start();
         k3po.awaitBarrier("ROUTED_INPUT");
@@ -91,9 +92,9 @@ public class MessageFormatIT
     @Test
     @Specification({
         "${route}/output/new/controller",
-        "${streams}/response.with.content.length/client/source",
-        "${streams}/response.with.content.length/client/target" })
-    public void shouldAcceptResponseWithContentLength() throws Exception
+        "${streams}/response.status.101.with.upgrade/client/source",
+        "${streams}/response.status.101.with.upgrade/client/target" })
+    public void shouldSwitchProtocolAfterUpgradeClient() throws Exception
     {
         k3po.start();
         k3po.awaitBarrier("ROUTED_OUTPUT");
@@ -104,9 +105,9 @@ public class MessageFormatIT
     @Test
     @Specification({
         "${route}/output/new/controller",
-        "${streams}/response.with.headers/client/source",
-        "${streams}/response.with.headers/client/target" })
-    public void shouldAcceptResponseWithHeaders() throws Exception
+        "${streams}/multiple.requests/client/source",
+        "${streams}/multiple.requests/client/target" })
+    public void shouldIssueMultipleRequests() throws Exception
     {
         k3po.start();
         k3po.awaitBarrier("ROUTED_OUTPUT");
@@ -117,27 +118,14 @@ public class MessageFormatIT
     @Test
     @Specification({
         "${route}/output/new/controller",
-        "${streams}/request.with.headers/client/source",
-        "${streams}/request.with.headers/client/target" })
-    public void shouldWriteRequestWithHeaders() throws Exception
+        "${streams}/multiple.requests.same.connection/client/source",
+        "${streams}/multiple.requests.same.connection/client/target" })
+    @Ignore("Not yet supported")
+    public void shouldIssueMultipleRequestsUsingConnectionPool() throws Exception
     {
         k3po.start();
         k3po.awaitBarrier("ROUTED_OUTPUT");
         k3po.notifyBarrier("ROUTED_INPUT");
         k3po.finish();
     }
-
-    @Test
-    @Specification({
-        "${route}/output/new/controller",
-        "${streams}/request.with.content.length/client/source",
-        "${streams}/request.with.content.length/client/target" })
-    public void shouldWriteRequestWithContentLength() throws Exception
-    {
-        k3po.start();
-        k3po.awaitBarrier("ROUTED_OUTPUT");
-        k3po.notifyBarrier("ROUTED_INPUT");
-        k3po.finish();
-    }
-
 }
